Fix builds on PowerBook G4
On the PowerBook G4, a 'uname -m' command returns "ppc"; however,
the architecture-specific files are located at arch/powerpc/... The
Makefile is modified to transform the ARCH variable appropriately,
and the appropriate little/big-endian mode is set. In addition, the
code in rtw_br_ext.c skips the call to the missing routine csum_ipv6_magic().
The latter change is temporary.

Although the driver can be built on the PowerPC architecture, it does not
yet work on that architecture.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>