Remove unused rtl8192d_PHY_ConfigRFWithParaFile, and fix so that scanning works with NetworkManager
There apparently was confusion between rtl8192d_PHY_ConfigRFWithHeaderFile and
rtl8192d_PHY_ConfigRFWithParaFile. The former is not an empty routine, but
the patch that said it was removing the latter actually substituted for the
former. This error likely happened when the patches for master were translated
to the kernel-version branch.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>