rtl8192du: Restore full speed to 5G
This patch fixes the remaining bug that was limiting the 5G rate.
Both RX and TX get in excess of 80 Mbps using netperf. In fact, the
actual rate could be higher as my netperf server is connected to the AP
with a 100 Mbps wired connection, and netperf only gets about 90 Mbps
over the wire.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>